backend/requirements.txt
Django==1.10.3
django-cors-headers==1.3.1
django-filter==0.15.3
djangorestframework==3.5.3
Markdown==2.6.7
wheel==0.24.0
